% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bonjour.R
\name{bonjour}
\alias{bonjour}
\title{Fonctions de demonstration, pour illustrer le fonctionnement des packages R.}
\usage{
bonjour(df)
}
\description{
Fonctions de demonstration, pour illustrer le fonctionnement des packages R.
}

